Merged in feature/qservice (pull request #13)
Feature/qservice

Added a minimal copy of Qservice, most code was copied as is with only changing the namespace. Seems to work, but needs testing. Qservice is running on port 5001 to not clash with Qserver. Not tested on the Pi yet.
Fixed strange warnings in test project. Needs to be a web app for some reason and have some explicit dependencies.
There’s a lot of stuff that can be cleaned up, but most of it is just a copy of the old Qservice calls. Will simplify and clean up later on.